fix(scripts): stop creating duplicate vector indexes, batch inserts, split chunking [T002315]
ensureSchema() prueft, ob bereits ein Vektor-Index existiert — und tat das
ueber `indexname LIKE '%ivfflat%'`. Diese Bedingung trifft NIE zu: Postgres
benennt einen namenlosen CREATE INDEX nach Tabelle und Spalte
(code_embeddings_embedding_idx, _idx1, _idx2 …), nie nach der
Zugriffsmethode. Jeder Lauf legte deshalb einen weiteren Index an; am
2026-07-27 lagen drei identische ivfflat-Indizes auf der Tabelle und
verdreifachten jeden INSERT. Die Pruefung laeuft jetzt ueber pg_index/pg_am.

Weitere Aenderungen aus der Messung desselben Tages (bge-m3 Q8_0, :8095):

- Multi-Row-Inserts statt eines Roundtrips pro Chunk. Gemessen ~30ms je
  Roundtrip, bei 18.549 Chunks also rund neun Minuten reine Wartezeit —
  waehrend das Einbetten derselben Menge ~110s braucht. Der Flaschenhals lag
  nie beim Modell. Ein Voll-Lauf sank damit auf 158 Sekunden.
- HNSW statt ivfflat: lists=100 war fuer den gewachsenen Bestand zu niedrig
  (Faustregel rows/1000). HNSW braucht keine Kalibrierung, bleibt beim
  Wachsen gueltig und baute die 18.549 Vektoren in 20 Sekunden auf.
- Optionale Parallelitaet ueber SCS_WORKERS. Default bleibt 1: der Regelfall
  ist der inkrementelle Lauf aus dem git-Hook, wo Nebenlaeufigkeit nichts
  bringt. SCS_WORKERS=4 (= die Slots des llama-servers) lohnt beim
  Voll-Neuaufbau.
- Dateien mit NUL-Bytes werden vor dem Insert aussortiert. Postgres laesst
  0x00 in TEXT strukturell nicht zu; betroffen sind zwei generierte Dateien
  unter scripts/docs-gen/, die bisher am INSERT scheiterten.

Dazu ein echter Split statt kosmetischem Zusammenziehen: index-repo.ts lag
mit 487 von 600 Zeilen bei 81 % der S1-Schwelle, die Aenderungen haetten auf
93 % gefuehrt. Die Chunking-Logik ist der Teil mit der klarsten Grenze —
pure Funktionen ueber Strings, ohne DB-, Netz- oder FS-Zugriff — und liegt
jetzt in scripts/lib/scs-chunking.ts. index-repo.ts re-exportiert sie, damit
bestehende Importpfade unveraendert bleiben (445 Zeilen, 74 %).
